add AllDaoTest test suite

+++ src/integration-test/java/tw/edu/ncu/cc/manage/dao/impl/ApiTokenDaoTest.java
@@ -81,6 +81,8 @@ public class ApiTokenDaoTest {
 	
 	@After
 	public void tearDown() {
-		this.clientDao.remove(client);
+		if (client != null) {
+			this.clientDao.remove(client);
+		}
 	}
 }
